feat: Add support for `mail_settings` message field (#118)
feat: Add support for `mail_settings` message field

Adds support for the `mail_settings` field of the
[mail-send](https://www.twilio.com/docs/sendgrid/api-reference/mail-send/mail-send#request-body)
API. This field includes the `sandbox_mode` setting, which allows
sending the API to SendGrid without actually sending the email to the
recipient(s), which is useful for testing.

Closes https://github.com/gsquire/sendgrid-rs/issues/103

Encode mutually exclusive bypass filters as an enum

When the `bypass_list_management` field is present, the other bypass
filters (`bypass_spam_management`, `bypass_bounce_management`, and
`bypass_unsubscribe_management`) are ignored. Model this as an enum in
the code.